Update rating format for naive SVD.

+++ src/model.hpp
@@ -10,7 +10,7 @@
 class Model {
     public:
         Model();
-        std::vector<std::vector<float> > ratings;
+        std::vector<std::vector<int> > ratings;
         virtual ~Model();
         void loadFresh(std::string fname);
         void loadSaved(std::string fname);
